refactor: 변수명 변경
